pyproject.toml
[project]
name = "gads-etl"
version = "0.1.0"
description = "Production grade Google Ads ETL pipeline template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-ads>=24.1.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.3",
    "pyyaml>=6.0.1",
    "dotenv>=0.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
gads-etl = "gads_etl.cli:app"

[tool.setuptools]
package-dir = {"" = "app/python-etl"}

[tool.setuptools.packages.find]
where = ["app/python-etl"]

[tool.pytest.ini_options]
addopts = "-ra -m 'not integration'"
testpaths = ["tests"]
markers = [
    "integration: tests that call external services and require credentials",
]
pythonpath = [
    "app/python-etl",
]
